+++ internal/component/azure/eventhubs/entity_management.go
