+++ internal/component/azure/eventhubs/track1_upgrade.go
